fix(RacingGame) : Remove RacingHandler assign responsibility to RacingGame
